add update functionality and break delete

+++ frontend/src/components/WorkoutDetails.js
@@ -1,21 +1,33 @@
 import useWorkoutsContext from "../hooks/useWorkoutsContext";
 import { format } from "date-fns";
 
-const WorkoutDetails = ({ workout }) => {
+const WorkoutDetails = ({ workout, editingState, updateEditingState }) => {
   const { dispatch } = useWorkoutsContext();
+  const isDisabled =
+    editingState._id !== "" && editingState._id !== workout._id;
 
-  const handleClick = async () => {
-    const response = await fetch("/api/workouts/" + workout._id, {
-      method: "DELETE",
-    });
-    const json = await response.json();
-    if (response.ok) {
-      dispatch({ type: "DELETE_WORKOUT", payload: json });
+  const handleDelete = async () => {
+    if (!isDisabled) {
+      const response = await fetch("/api/workouts/" + workout._id, {
+        method: "DELETE",
+      });
+      const json = await response.json();
+      if (response.ok) {
+        dispatch({ type: "DELETE_WORKOUT", payload: json });
+      }
+      updateEditingState(null);
     }
   };
-
+  const handleEdit = () => !isDisabled && updateEditingState(workout);
+  const handleCancel = () =>
+    updateEditingState({
+      _id: "",
+      title: "",
+      load: "",
+      reps: "",
+    });
   return (
-    <div className="workout-details">
+    <div className={"workout-details " + (isDisabled ? "shaded" : "")}>
       <h4>{workout.title}</h4>
       <p>
         <strong>Load (kg): </strong>
@@ -27,9 +39,21 @@ const WorkoutDetails = ({ workout }) => {
       </p>
       <p>{format(new Date(workout.createdAt), "d-MMM-yyyy")}</p>
 
-      <span onClick={handleClick} className="material-symbols-outlined">
+      <span onClick={handleDelete} className="material-symbols-outlined delete">
         Delete
       </span>
+      {editingState._id === workout._id ? (
+        <span
+          onClick={handleCancel}
+          className="material-symbols-outlined close"
+        >
+          Close
+        </span>
+      ) : (
+        <span onClick={handleEdit} className="material-symbols-outlined edit">
+          Edit
+        </span>
+      )}
     </div>
   );
 };
